fixed few symbol remapping issues

+++ core/src/main/java/io/questdb/std/Vect.java
@@ -230,7 +230,7 @@ public final class Vect {
 
     public static native int mergeShuffleStringColumnFromManyAddresses(int mergeIndexEncodingSegmentBytes, int dataLengthBytes, long primaryAddressList, long secondaryAddressList, long outPrimaryAddress, long outSecondaryAddress, long mergeIndex, long rowCount, long destVarOffset);
 
-    public static native int mergeShuffleSymbolColumnFromManyAddresses(int mergeIndexEncodingSegmentBytes, long srcAddresses, long dstAddress, long mergeIndex, long rowCount, long txnInfo, long txnCount, long symbolMapAddress, long symbolMapSize);
+    public static native long mergeShuffleSymbolColumnFromManyAddresses(int mergeIndexEncodingSegmentBytes, long srcAddresses, long dstAddress, long mergeIndex, long rowCount, long txnInfo, long txnCount, long symbolMapAddress, long symbolMapSize);
 
     public static native int mergeShuffleVarcharColumnFromManyAddresses(int mergeIndexEncodingSegmentBytes, long primaryAddressList, long secondaryAddressList, long outPrimaryAddress, long outSecondaryAddress, long mergeIndex, long rowCount, long destVarOffset);
